+++ crates/fluvio-sc-schema/src/topic/mod.rs
@@ -1,5 +1,44 @@
 pub use fluvio_controlplane_metadata::topic::*;
+pub mod validate {
+    /// Ensure a topic can be created with a given name.
+    /// Topics name can only be formed by lowercase alphanumeric elements and hyphens.
+    /// They should start and finish with an alphanumeric character.
+    pub fn valid_topic_name(name: &str) -> bool {
+        name.chars()
+            .all(|ch| ch.is_ascii_lowercase() || ch.is_ascii_digit() || ch == '-')
+            && !name.ends_with('-')
+            && !name.starts_with('-')
+    }
 
+    #[cfg(test)]
+    mod tests {
+        use crate::topic::validate::valid_topic_name;
+
+        #[test]
+        fn reject_topics_with_spaces() {
+            assert_eq!(valid_topic_name("hello world"), false);
+        }
+
+        #[test]
+        fn reject_topics_with_uppercase() {
+            assert_eq!(valid_topic_name("helloWorld"), false);
+        }
+
+        #[test]
+        fn reject_topics_with_underscore() {
+            assert_eq!(valid_topic_name("hello_world"), false);
+        }
+
+        #[test]
+        fn valid_topic() {
+            assert!(valid_topic_name("hello-world"));
+        }
+        #[test]
+        fn reject_topics_that_start_with_hyphen() {
+            assert_eq!(valid_topic_name("-helloworld"), false);
+        }
+    }
+}
 mod convert {
 
     use std::convert::TryInto;